% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_calculate_projections.R
\name{adapted_calculate_projections}
\alias{adapted_calculate_projections}
\title{adapted_calculate_projections}
\usage{
adapted_calculate_projections(
  parameters,
  initial_condition,
  t_window,
  time_step_len = 0.01,
  model = "GLV",
  derivative = EEMtoolbox::derivative_func,
  scaled = FALSE,
  species_names = NA,
  mode = "recruitment",
  init_intervention_amount = 0,
  init_intervention_timepoints = NA,
  sustain_intervention_amount = 0,
  sustain_intervention_timepoints = NA,
  sustain_intervention_threshold = NA,
  intro_species_index = 1,
  multiplier = 1,
  extinction_threshold = 0
)
}
\arguments{
\item{parameters}{list like object of ensemble of parameters (outputs of  \link[EEMtoolbox]{EEM})}

\item{initial_condition}{vector of initial species abundances. If parameter scaled is TRUE, the parameter initial_condition should be scaled to steady state}

\item{t_window}{time window to solve ODE}

\item{time_step_len}{length of each time step, default = 0.01}

\item{model}{model representing species interactions. Default "GLV" (Generalized Lotka Volterra). options include "Bimler-Baker" and "Gompertz".}

\item{derivative}{derivative function. Default \link[EEMtoolbox]{derivative_func}}

\item{scaled}{Boolean indicating if projections should be scaled to steady state. If true, the parameter initial_condition should be scaled too. Default FALSE}

\item{species_names}{vector of strings for names of species. If NA plots only display species index number, . Default NA.}

\item{mode}{"recruitment" or "removal". Default "recruitment"}

\item{init_intervention_amount}{vector with the number of individuals in initial interventions for each introduced species, default = 0}

\item{init_intervention_timepoints}{vector or list (if multiple introduced species) with the initial intervention timepoints for each introduced species, default = NA}

\item{sustain_intervention_amount}{vector with the number of individuals in subsequent interventions for each introduced species, default = 0}

\item{sustain_intervention_timepoints}{vector or list (if multiple introduced species) with the subsequent intervention timepoints for each introduced species, default = NA}

\item{sustain_intervention_threshold}{vector or list (if multiple introduced species) with the abundance threshold where subsequent interventions stop for each introduced species, default = NA}

\item{intro_species_index}{vector indicating which species in the index is/are the introduced one(s), default = 1}

\item{multiplier}{multiplier for the initial condition, default = 1}

\item{extinction_threshold}{threshold below which species are considered extinct, default = 0}
}
\value{
dataset of species abundances over time
}
\description{
Solves ODEs and plot solutions
}
\examples{
library(EEMtoolbox)
output <- EEMtoolbox::EEM(matrix(c(-1,-1,1,-1),ncol =2), n_ensemble = 2)
adapted_calculate_projections(output,
                              initial_condition = rep(1,2),
                              t_window = c(0,1),
                              mode = "recruitment",
                              init_intervention_amount = 2,
                              init_intervention_timepoints = c(1, 2))

}
\author{
Maude Vernet
}
