% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_ode_solve.R
\name{adapted_ode_solve}
\alias{adapted_ode_solve}
\title{adapted_ode_solve}
\usage{
adapted_ode_solve(
  initial_condition,
  interaction_matrix_value,
  growth_rate,
  t_window,
  time_step_len,
  model = "GLV",
  derivative = EEMtoolbox::derivative_func,
  recruitment_pars,
  recruitment_event,
  recruitment_times,
  extinction_threshold
)
}
\arguments{
\item{initial_condition}{A numeric vector representing the initial conditions
of the system.}

\item{interaction_matrix_value}{A matrix representing the interaction
coefficients between species.}

\item{growth_rate}{A numeric vector representing the growth rates of the
species.}

\item{t_window}{A numeric vector of length 2 representing the time window for
the simulation (start and end times).}

\item{time_step_len}{A numeric value representing the length of each time
step in the simulation.}

\item{model}{A character string specifying the model type (default is "GLV"
for Generalized Lotka-Volterra).}

\item{derivative}{A function that computes the derivatives of the system
(default is \code{EEMtoolbox::derivative_func}).}

\item{recruitment_pars}{A list of parameters related to recruitment events.}

\item{recruitment_event}{A function that defines the recruitment event.}

\item{recruitment_times}{A numeric vector specifying the times at which
recruitment events occur.}

\item{extinction_threshold}{A numeric value representing the threshold below
which species are considered extinct.}
}
\value{
A data frame containing the results of the ODE solution, including
time steps and species populations.
}
\description{
This function is a wrapper for the \code{deSolve::ode} function, which solves
ordinary differential equations (ODEs) for ecological models.
It allows for the inclusion of recruitment events and extinction thresholds,
making it suitable for ecological modeling scenarios.
It is adapted from the \code{EEMtoolbox::ode_solve} function.
}
\author{
Maude Vernet
}
