% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species_names.R
\name{add_species_names}
\alias{add_species_names}
\title{add_species_names}
\usage{
add_species_names(parameter, species_names)
}
\arguments{
\item{parameter}{a set of parameters obtained with \code{EEMtoolbox::EEM()}.}

\item{species_names}{Character vector of length p. The new names to assign.}
}
\value{
The same object \code{parameter} but with the names set to
\code{species_names} on the interaction matrix.
}
\description{
Add species names to a numeric vector or matrix
This helper takes an ´EEM´ object and adds species names to the
interaction matrix.
}
\examples{
output <- EEMtoolbox::EEM(matrix(c(-1, -1, 1, -1), ncol = 2), n_ensemble = 2)
withnames <- add_species_names(output,
                               c("sp1", "sp2"))
}
\author{
Maude Vernet
}
