% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eq.R
\name{extract_eq}
\alias{extract_eq}
\title{extract_eq}
\usage{
extract_eq(EEM, model = "GLV")
}
\arguments{
\item{EEM}{A list of EEM objects}

\item{model}{The model used for the EEM objects. Default is "GLV".}
}
\value{
A list of equilibrium values for each EEM object
}
\description{
Extract equilibrium values from parameter sets of an EEM object
}
\examples{
out <- EEMtoolbox::EEM(matrix(c(-1, -1, 1, -1), ncol = 2), n_ensemble = 2)
equilibrium_values <- extract_eq(out)
}
\author{
Maude Vernet
}
