% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_introductions.R
\name{merge_introductions}
\alias{merge_introductions}
\title{Merge multiple single-species introductions into one system}
\usage{
merge_introductions(EEM_intros, sign_interaction_intros = 0, mode = "updated")
}
\arguments{
\item{EEM_intros}{a pairlist of objects obtained with
add_introduced_species()}

\item{sign_interaction_intros}{one number or a matrix of signs for the
interactions between introduced species and native species (intraspecies
interactions to be marked as NA).}

\item{mode}{"updated" or "recycled". If "updated", the interaction signs are
updated for each projection. If "recycled", the interaction signs are
recycled in all projections}
}
\value{
a list of EEM objects
}
\description{
This function merges multiple single-species introductions into one
ecosystem.
}
\examples{
out1 <- EEMtoolbox::EEM(matrix(c(-1, -1, 1, -1), ncol = 2), n_ensemble = 2)
out1_intro <- add_introduced_species(out1,
                                     introduced_self_sign = -1,
                                     introduced_row_signs = c(1, 1),
                                     introduced_col_signs = c(-1, -1),
                                     introduced_k = 5)
out2 <- EEMtoolbox::EEM(matrix(c(-1, -1, 1, -1), ncol = 2), n_ensemble = 2)
out2_intro <- add_introduced_species(out2,
                                     introduced_self_sign = -1,
                                     introduced_row_signs = c(1, 1),
                                     introduced_col_signs = c(-1, -1),
                                     introduced_k = 10)
merged <- merge_introductions(pairlist(out1_intro, out2_intro),
                              sign_interaction_intros = 0,
                              mode = "updated")
}
\author{
Maude Vernet
}
