% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_discrepancy_continuous_sum.R
\name{adapted_discrepancy_continuous_sum}
\alias{adapted_discrepancy_continuous_sum}
\title{Adapted disc_func for EEM}
\usage{
adapted_discrepancy_continuous_sum(
  data,
  target_lower = NULL,
  target_upper = NULL
)
}
\arguments{
\item{data}{vector of the equilibrium points followed by the real eigenvalues of the jacobian matrix, as returned by \link[EEMtoolbox]{summarise_ecosystem_features}}

\item{target_lower}{Numeric vector of length p. If provided, these values will
serve as the minimum allowable equilibrium abundances for each of the p species.}

\item{target_upper}{Numeric vector of length p. If provided, these values will
serve as the maximum allowable equilibrium abundances for each of the p species.}
}
\value{
A list with the same elements as \code{EEMtoolbox::dicrepancy_continuous_sum()}
}
\description{
A wrapper around \code{EEMtoolbox::dicrepancy_continuous_sum()} that
}
\details{
This is exactly the same logic as \code{dicrepancy_continuous_sum()} in EEMtoolbox,
except that the standard EEMtoolbox version does \emph{not} check for user‐specified
equilibrium bounds.

\verb{important notice}: numbers in \code{target_lower} and \code{target_upper} are to be
inputed directly, and not inside a variable. Otherwise, they are not found by
the function.
}
\examples{
out <- EEMtoolbox::EEM(
             interaction_matrix = matrix(c(-1, -1, 1, -1), ncol = 2),
             n_ensemble = 2,
             disc_func = function(data) {
             EEMtoolbox::adapted_discrepancy_continuous_sum(
             data,
             target_lower = rep(1,2),
             target_upper = rep(20,2))})

}
\seealso{
\code{EEMtoolbox::dicrepancy_continuous_sum()} (original function)
}
\author{
Maude Vernet
}
